feat: add CutListClass interface and implement conversion to CutList
